requirements.txt
aws-cdk.core==1.134.0
aws-cdk.aws_ec2==1.134.0
aws-cdk.aws_s3==1.134.0
aws-cdk.aws_ecs==1.134.0
aws-cdk.aws_rds==1.134.0
aws-cdk.aws_iam==1.134.0
aws-cdk.aws_secretsmanager==1.134.0
aws-cdk.aws_ecs_patterns==1.134.0
aws-cdk.aws_elasticloadbalancingv2==1.134.0
aws-cdk.aws_elasticloadbalancingv2_targets==1.134.0
aws-cdk.aws_route53==1.134.0
aws-cdk.aws_certificatemanager==1.134.0